refactor: move input types to types file
